FUNCIONAL CARAMBAAAAA

+++ projetotech/src/main/java/com/gympoison/projeto_tech/repository/Auditoria_de_precoRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Auditoria_de_precoRepository extends JpaRepository<Auditoria_de_preco, Integer> {
+public interface Auditoria_de_precoRepository extends JpaRepository<Auditoria_de_preco, Long> {
 }
+
